pyproject.toml
[tool.poetry]
name = "hackathon-pictionnary-llm"
version = "0.1.0"
description = ""
authors = ["hymaia"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
tk = "^0.1.0"
langchain = "^0.2.9"
pillow = "^10.4.0"
httpx = "^0.27.0"
langchain-openai = "^0.1.17"
streamlit-drawable-canvas = "^0.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
paint = "src.main:main"
